Moved runner.js to exercises, didn't quite belong at root
