+++ exercises/basic_every_some/exercise.js
@@ -1,7 +1,7 @@
 "use strict"
 
 var random = require('../randomizer')
-var runner = require('../../runner')
+var runner = require('../runner')
 
 function makeUser() {
   return {
