+++ exercises/basic_filter/exercise.js
@@ -2,7 +2,7 @@
 
 var loremIpsum = require('lorem-ipsum')
 var random = require('../randomizer')
-var runner = require('../../runner')
+var runner = require('../runner')
 
 var input = random.arrayOf(10, 30, function() {
   return { message: loremIpsum() }
