+++ exercises/hello_world/exercise.js
@@ -1,5 +1,5 @@
 "use strict"
 
-var runner = require('../../runner')
+var runner = require('../runner')
 
 module.exports = runner(require('lorem-ipsum')())
